rootfs/rootfs/opt/zoho-calendar/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dateutil==2.8.2
paho-mqtt==1.6.1
schedule==1.2.1
